==> notebooks/01_textbook-question-EDA.ipynb <==
# 01. Textbook Problem Extraction and EDA

Our goal for this project is to be able to balance a chemical equation that may or may not be presented in plain English. This will involve several steps:

- parsing text
- 

In this notebook we will use `tika` to extract back-of-the-chapter problems from textbook `.pdf` files. As an initial comparison, we'll limit our scope to the following topics: `chemical reactions` vs. `quantum mechanics`. To make the processing a little easier, the pages containing the problems have been extracted from the main textbook files using `Preview` (any `pdf` editor should work).

import os                               
import re                   
import time                 # to stall requests (just in case)

import pandas as pd 
import chemdataextractor as cde     # chemistry parser

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize, sent_tokenize

import tika                 # to initiate tika server
from tika import parser     # the specific parser method      

pd.set_option('display.max_colwidth', 0)    # no max column width
pd.set_option('display.max_rows', 150)      

filepaths = []

for file in os.listdir('../data/textbook-problems/'):
    if 'pdf' in file:
        filepaths.append('../data/textbook-problems/'+file)
        
filepaths

## Examine the Text

def grab_text(file, sleep=0, counter=0):
    
    if counter == 2:        # so we stop the recursive function
        pass

    # grab the raw text using parser.from_file()
    raw = parser.from_file(file)    
    status = raw['status']          # returns the status code from tika server
    
    # if things go well, return the raw text
    if status == 200:               
        print(f"'{file}' successfully opened!")
        return raw['content']
    
    # if things don't go well, pause for five seconds and try again
    # we might not need this code, but it's useful for other server calls
    else:                           
        print(f'! ! ! ! error code {status} ! ! ! !')
        print(f'! ! ! ! trying again ! ! ! !')
        
        time.sleep(5)
        counter += 1
        
        # repeats grab_text up to twice
        return grab_text(file, counter=counter) 

We will be using `regex` to extract the problems from each file. Different textbooks may label their questions differently, so we will need to examine how each textbook is formatted so we only grab the relevant information.

test = grab_text('../data/textbook-problems/bauer-7.pdf')
test

If we examine how the text is formatted in this text, we see that each new problem is formatted with a line break, the chapter number, a period, and the problem number:

`\n 7.4 List three types of electromagnetic radiation that have \n\nhigher frequencies than visible light.\n 7.5 Draw a picture of two waves, one with twice the \n\nfrequency of the other. Label the wave with the \nhigher frequency.`

Notice how there are also line breaks in the problem itself. These shouldn't be included in the cleaned text.

Every textbook has their own way of formatting the problems.

## Clean the Text

The following function attempts to catch all of the variations.

# https://stackoverflow.com/questions/44333462/
def make_problems(document):
    
    # '\n 7.16 ' => 'PPRROOBBLLEEMM'
    clean = re.sub('\n\s[0-9]+\.[0-9]+\s', r'PPRROOBBLLEEMM', document)
    
    # '  16. ' => 'PPRROOBBLLEEMM'
    clean = re.sub('\s\s[0-9]+\.\s', r'PPRROOBBLLEEMM', clean)
    
    # '\n\n 16. ' => 'PPRROOBBLLEEMM'
    clean = re.sub('\n\n\s[0-9]+\. ', r'PPRROOBBLLEEMM', clean)
    
    # '\n16. ' => 'PPRROOBBLLEEMM'
    clean = re.sub('\n([0-9]+)\.\s', r'PPRROOBBLLEEMM', clean)

    # '\na. ' => ' '
    # joins sub-problems with the main problem
    clean = re.sub('\n([a-z]+)\.\s', ' ', clean)
    
    clean = re.sub('\-\n', '', clean)       # join hyphenated words
    clean = re.sub('\n', ' ', clean)        # treat line breaks as spaces
    clean = re.sub('\t', ' ', clean)        # treat tabs as spaces
    clean = re.sub('\s\s', ' ', clean)      # treat double-spaces as singles

    clean = re.split('PPRROOBBLLEEMM', clean)   # split by problem
    
    return clean

make_problems(test)

PROBLEMS = []

for file in filepaths: 
    name = file[26:-4]          # file[:26] = '../data/textbook-problems/'
    text = make_problems(grab_text(file))
    
    for i, problem in enumerate(text):
        problem_dict = {}
        problem_dict['filepath'] = name
        problem_dict['number'] = i
        problem_dict['text'] = problem.strip()
        PROBLEMS.append(problem_dict)

# turn list of dictionaries into DataFrame
df = pd.DataFrame(PROBLEMS, columns=['filepath', 'number', 'text'])

print(df.shape)
df.head()

### Tokenizing and Removing Stopwords

It may be useful to tokenize the problems. We will use two methods:

- `NLTK` `word_tokenize()`, removing English stopwords
- `ChemDataExtractor` chemical entity mentions

stop_words = stopwords.words('english')
stop_words += ['copyright',
               'cengage',
               'pearson',
               'learning',
               'may',
               'copied',
               'scanned',
               'duplicated',
               'chapter',
               'practice',
               'problem',
               'exercise',
               'review',
               'question',
               'figure']
stop_words

def remove_stops(doc):
    doc = word_tokenize(doc)
    doc = [w.lower() for w in doc if not w in stop_words] 
    doc = [w for w in doc if w.isalpha()]
    return doc

df['txt'] = df['text'].apply(remove_stops)
df.head()

df['text'] = df['text'].apply(sent_tokenize)
df.head()

r = cde.doc.Paragraph(df.loc[0, 'text'])
s = [i for i in r.pos_tagged_tokens]
s

s = r.pos_tagged_tokens

s[0][0][0]

t = []
for w in df.loc[0, 'txt']:
    if w not in s:
        t.append(w)
    else:
        t.append('CPD')

t

df['filepath'].value_counts()

## Labeling Problems

df['txt'] = df['text'].apply(remove_stops)
df.reset_index(drop=True, inplace=True)

def check(df):
    print(df.shape)
    return df.head()

# code adapted from Dae H, Sophia A, Sonam T : LA

def label_problem(df, start=0, problem_col='text'):

    balancing = []
    e_config = []
    
    if start > 0:
        for i in range(start):
            balancing.append(df.loc[i, problem_col]['balancing'])
            e_config.append(df.loc[i, problem_col]['e_config'])
    else:
        pass

    for i in range(start, len(df)):
        print('=====================')
        print()
        print(f'{i+1}/{len(df)}')
        print(df.loc[i, problem_col])
        print()
        bal = input('answer specifically asks for balanced equation: 1 -- ')
        if bal == "stop":
            for n in range(len(df) - (i)):
                balancing.append(999)
            
            dataframe['balancing'] = balancing
            print('function stopped')
            print(f'stopped at index {i}.')
            break
        else:
            balancing.append(bal)
        print()
        ec = input('answer specifically asks for electron configuration: 1 -- ')
        if ec == "stop":
            for n in range(len(df) - (i)):
                e_config.append(999)
            
            dataframe['e_config'] = e_config
            print('function stopped')
            print(f'stopped at index {i}.')
            break
        else:
            e_config.append(ec)
    
    df['balancing'] = balancing
    df['e_config'] = e_config
    
    return df

# df = label_problem(df)

df.dropna(axis=0, inplace=True)

df.to_csv('../data/textbook-problems.csv', index=False)
