==> notebooks/02_chemdataextractor-and-stanfordnlp_problem-parsing.ipynb <==
import chemdataextractor as cde
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
import pickle
import numpy as np
import pandas as pd
import stanfordnlp
import spacy
from spacy.matcher import Matcher

nlp = spacy.load('en_core_web_lg')

import tools

tools.formula_state_separator('NaCl(aq)')

nlp = spacy.load('en_core_web_lg')
matcher = Matcher(nlp.vocab)
text = 'methane is reacted with oxygen to form carbon dioxide and water vapor.'
doc = nlp(text)
sents = list(doc.sents)
print("Number of Sentences = ",len(sents))
for sent in doc.sents:
    for token in sent:
        print(token.dep_,token.tag_, end = " ")
    print()
passive_rule = [{'DEP':'nsubjpass'},{'DEP':'aux','OP':'*'},{'DEP':'auxpass'},{'TAG':'VBN'}]
matcher.add('Passive',None,passive_rule)
matches = matcher(doc)
print(len(matches))

## Types of balancing questions:

- Reaction without product. Problem statement.  
    Sodium metal reacts with chlorine gas in a combination reaction. Write a balanced equation to describe this reaction.
- Reaction with product. Problem statement.  
    Sodium metal reacts with chlorine gas to form sodium chloride.  
    Sodium metal and chlorine gas react to form sodium chloride.
- Problem statement: equation.  
    Balance the following equation: Na(s) + Cl2(g) -> NaCl(s)
- Problem statement, reaction.  
    Write a balanced chemical equation to describe the reaction between sodium metal and chlorine gas.  
    Complete and balance the equations for each of the following reactions.

We have to download the appropriate language model for our neural pipeline.

stanfordnlp.download('en')

nlp = stanfordnlp.Pipeline()
doc = nlp('methane and oxygen form carbon dioxide and water.')
print(*[f'text: {word.text+" "}\tlemma: {word.lemma}\tupos: {word.upos}\txpos: {word.xpos}' for sent in doc.sentences for word in sent.words], sep='\n')ljust

nlp = stanfordnlp.Pipeline(processors='tokenize,mwt,pos,lemma,depparse', lang='en')
doc = nlp('methane and oxygen form carbon dioxide and water.')

doc.sentences[0].words[0].dependency_relation

print([f'text: {word.text}\talemma: {word.lemma}' for sent in doc.sentences for word in sent.words], sep='\n')

doc = nlp('oxygen and methane react to form carbon dioxide and water.')
print(*[f'text: {word.text+" "}\tlemma: {word.lemma}\txpos: {word.xpos}\tdeprel: {word.dependency_relation}' for sent in doc.sentences for word in sent.words], sep='\n')

doc = nlp('carbon dioxide forms when methane combusts in oxygen.')
print(*[f'text: {word.text+" "}\tlemma: {word.lemma}\txpos: {word.xpos}\tdeprel: {word.dependency_relation}' for sent in doc.sentences for word in sent.words], sep='\n')

nlp = stanfordnlp.Pipeline(processors='tokenize,mwt,pos,lemma,depparse', lang='en')

reaction_phrases = [
    'methane reacts with oxygen to form carbon dioxide and water vapor.',
    'CH4 reacts with O2 to form CO2 and H2O.',
    'methane and oxygen form carbon dioxide and water.',
    'oxygen and methane react to form carbon dioxide and water.',
    'methane and oxygen react in a combustion reaction.',
    'methane combusts in the presence of oxygen.',
    'write a balanced chemical equation to describe the combustion of methane and oxygen.',
    'carbon dioxide is formed when methane combusts in oxygen.'
]

for r in reaction_phrases:
    doc = nlp(r)
    r2 = cde.doc.Paragraph(r)
    print(*[f'text: {word.text+" "}\txpos: {word.xpos}\tdeprel: {word.dependency_relation}' for sent in doc.sentences for word in sent.words], sep='\n')
    print(r2.cems)
    print('===========')
    print()

r2.cems

z = Problem(sample)

z.question

print(z.inference)

def remove_stops(doc, i=0, progress=None):
    if progress:
        progress(i)
    doc = word_tokenize(doc)  # Split into words.
    doc = [w.lower() for w in doc if not w in stop_words]  # Remove stopwords.
    doc = [w for w in doc if w.isalpha()]  # Remove numbers and punctuation.
    return doc

balance_phrases = [
    'balance the following equations',
    'balance the following chemical equation',
    'complete and balance the equation',
    'use coefficients to balance each equation',
    'write a balanced chemical equation',
    'write a balanced molecular equation',
    'what is the balanced chemical equation',
    'when the equation above is balanced with lowest whole-number coefficients, the coefficient for is'
]

balanced = [remove_stops(sent) for sent in balance_phrases]
balanced

model = pickle.load(open('../model/model.p', 'rb'))
w2v = pickle.load(open('../model/w2v.p', 'rb'))

word_vectors = model.wv

probs = pd.read_csv('../data/textbook-problems.csv')

probs.head()

for i in probs.index:
    distlist = [word_vectors.wmdistance(probs.loc[i, 'txt'], ' '.join(j)) for j in balanced]
    probs.loc[i, 'distance'] = np.mean(distlist)
probs.sort_values(by='distance')[:50]
