==> notebooks/03_CHNOSZ_thermodynamics-and-stoichiometry.ipynb <==
import os                               
import re                   
import itertools
import pickle

import numpy as np
import pandas as pd 

from chempy import Substance

from tika import parser     # the specific parser method 

## Import thermodynamic values

These values were taken from the following sources:
- [CHNOSZ](http://chnosz.net/vignettes/obigt.html), a chemistry and materials science package for [R](https://www.r-project.org/about.html). The dataframes were exported to csv files as-is.
- [Principles of Modern Chemistry, 8th ed.](https://www.amazon.com/Principles-Modern-Chemistry-David-Oxtoby-ebook/dp/B00UGDPNFI) by Oxtoby et al.

### CHNOSZ

chnosz = pd.read_csv('../data/external/thermo/chnosz_thermo.csv')

chnosz.columns

chnosz = chnosz[['name', 'abbrv', 'formula', 'E_units', 'state', 'G', 'H', 'S', 'Cp']]
chnosz.head()

chnosz.shape

chnosz = chnosz[chnosz['name'].str[0] != '[']
chnosz.shape

chnosz['E_units'].value_counts()

# convert to joules
energy_columns = ['G', 'H', 'S', 'Cp']

for col in energy_columns:
    chnosz.loc[chnosz['E_units'] == 'cal', col] *= 4.184
    
chnosz.drop(columns='E_units', inplace=True)
chnosz.head()

chnosz.at[0, 'G'] = -237180
chnosz.at[0, 'H'] = -285830
chnosz.at[0, 'S'] = 69.91
chnosz.at[0, 'Cp'] = 75.29

# just checking to make sure appropriate values are in joules

print(chnosz.loc[4, 'G'])
print(chnosz.loc[779, 'G'])

chnosz['state'].value_counts()

value_dict = {
    'aq': '(aq)',
    'cr': '(s)',
    'liq': '(l)',
    'gas': '(g)',
    'cr2': '(s, II)',
    'cr3': '(s, III)',
    'cr4': '(s, IV)',
    'cr5': '(s, V)',
    'cr6': '(s, VI)',
    'cr7': '(s, VII)',
    'cr8': '(s, VIII)',
    'cr9': '(s, IX)',
    'g': '(g)',
}

chnosz['state'].replace(value_dict, inplace=True)

chnosz['formula'] = chnosz['formula'].astype(str) + chnosz['state'].astype(str)
chnosz.drop(columns='state', inplace=True)
chnosz.head()

chnosz.at[1, 'formula'] = 'e-(aq)'

chnosz.isna().sum()

chnosz = chnosz.dropna(subset=['G'])
chnosz.shape

### Oxtoby

def get_text(file, sleep=0, counter=0):
    if counter == 2:        # so we stop the recursive function
        pass
    # grab the raw text using parser.from_file()
    raw = parser.from_file(file)
    status = raw['status']          # returns the status code from tika server
    # if things go well, return the raw text
    if status == 200:
        print(f"'{file}' successfully opened!")
        return raw['content']
    # if things don't go well, pause for five seconds and try again
    # we might not need this code, but it's useful for other server calls
    else:
        print(f'! ! ! ! error code {status} ! ! ! !')
        print(f'! ! ! ! trying again ! ! ! !')
        time.sleep(5)
        counter += 1
        # repeats grab_text up to twice
        return get_text(file, counter=counter)

oxtobya = get_text('../data/external/thermo/oxtoby8a.pdf')
oxtobyb = get_text('../data/external/thermo/oxtoby8b.pdf')

clean = re.sub('—', 'nan', oxtobya)
clean = re.sub('\\ue02c', 'l', clean)
clean = re.sub(r'\n\n[I]*\s*', r'RROOWW', clean)
clean = re.sub(r'([\d]+)\n([A-Z]+)', r'\1RROOWW\2', clean)
clean = re.sub(r'(nan)\s*\n([A-Z]+)', r'\1RROOWW\2', clean)
clean = re.sub('\\ue031', '+', clean)
clean = re.sub('([\d])\+\(', r'+\1(', clean)
clean = re.sub('\\ue032', '-', clean)
clean = re.sub('([\d])\-\(', r'-\1(', clean)
clean = re.sub('\(([aqslg]+)\,\s([\w]+)', r'(\1,\2', clean)
clean = re.sub('[\s]+', ' ', clean)
clean = re.sub(' mol-1', '', clean)
# clean = re.sub('Substance.*mol', 'HHEEAADDEERR', clean)
clean = re.split('RROOWW', clean)
clean = [re.split(' ', r) for r in clean]
clean = [r for r in clean if len(r) == 5]
oxtobya_clean = clean

clean = re.sub('—', 'nan', oxtobyb)
clean = re.sub('\\ue02c', 'l', clean)
clean = re.sub(r'\n\n[I]*\s*', r'RROOWW', clean)
clean = re.sub(r'([\d]+)\n([A-Z]+)', r'\1RROOWW\2', clean)
clean = re.sub(r'(nan)\s*\n([A-Z]+)', r'\1RROOWW\2', clean)
clean = re.sub('\\ue031', '+', clean)
clean = re.sub('([\d])\+\(', r'+\1(', clean)
clean = re.sub('\\ue032', '-', clean)
clean = re.sub('([\d])\-\(', r'-\1(', clean)
clean = re.sub('\(([aqslg]+)\,\s([\w]+)', r'(\1,\2', clean)
clean = re.sub('[\s]+', ' ', clean)
clean = re.sub(' mol-1', '', clean)
# clean = re.sub('Substance.*mol', 'HHEEAADDEERR', clean)
clean = re.split('RROOWW', clean)
clean = [re.split(' ', r) for r in clean]
clean = [r for r in clean if len(r) == 5]
oxtobyb_clean = clean

oxtoby = oxtobya_clean + oxtobyb_clean
oxtoby[:20]

oxtoby[-20:]

oxtoby_df = pd.DataFrame(oxtoby, columns=['formula', 'H', 'S', 'G', 'Cp'])
oxtoby_df.head()

oxtoby_df.shape

energy_columns = ['G', 'H', 'S', 'Cp']

for col in energy_columns:
    oxtoby_df[col] = oxtoby_df[col].astype(float)
    
oxtoby_df.head()

oxtoby_df.loc[:, 'H'] *= 1000
oxtoby_df.loc[:, 'G'] *= 1000
oxtoby_df.head()

oxtoby_df = oxtoby_df[['formula', 'G', 'H', 'S', 'Cp']]
oxtoby_df.head()

oxtoby_df.isna().sum()

oxtoby_df = oxtoby_df.dropna(subset=['G'])
oxtoby_df.shape

## Combining Tables

thermo_df = pd.concat([chnosz, oxtoby_df], ignore_index=True).fillna(np.NaN)
thermo_df.head()

thermo_df = thermo_df[['formula', 'abbrv', 'name', 'G', 'H', 'S', 'Cp']]
thermo_df.tail()

thermo_df.shape

thermo_df.drop_duplicates(subset='formula', keep='last', inplace=True)
thermo_df.reset_index(drop=True, inplace=True)
thermo_df.shape

thermo_df[thermo_df['formula'].str.contains('\.')].index

thermo_df = thermo_df.drop(index=thermo_df[thermo_df['formula'].str.contains('\.')].index)
thermo_df.reset_index(drop=True, inplace=True)
thermo_df.shape

def formula_state_separator(formula, keep_state=False):
    '''
    Separates the state from a formula string.
    
    --Parameters--
    formula:        str
        a string of a single chemical formula
    
    --Output--
    tuple (str)
        
    --Examples--
    >>> formula_state_separator('NaCl(aq)')
    ('NaCl', 'aq')
    
    >>> formula_state_separator('NaCl')
    'NaCl'
    '''
    try:
        regex = re.search('(?<=\()[aglsq]+', formula)
        formula = formula[:regex.start() - 1]
        if keep_state:
            state = regex.group(0)
            return formula, state
        else:
            return formula
    except:
        return formula_state_separator

formulas = [formula_state_separator(f) for f in thermo_df['formula'].astype(str)]
formulas[:5]

new_list = []
for a, f in zip(list(thermo_df['abbrv']), formulas):
    if pd.isna(a):
        new_list.append(f)
    else:
        new_list.append(a)
new_list[:10]

thermo_df['abbrv'] = new_list

thermo_df.head()

thermo_df.tail()

masses = [Substance.from_formula(f).mass for f in thermo_df['formula'].astype(str)]
masses[:5]

thermo_df['mass'] = masses

pickle.dump(thermo_df, open('../data/processed/thermo_df.p', 'wb'))

## replicate CHNOSZ stoichiometry csv

Originally, I had imported the `CHNOSZ` stoichiometry csv. However, I found the need to update it with whatever new entries were present in `thermo`.

The stoichiometry csv is formatted so that each formula is count-vectorized by element.

stoich_og = pd.read_csv('../data/external/thermo/chnosz_stoich.csv')
stoich_og.head()

For my workflow, we will rename the columns by atomic number and have charge occupy column `0` to be in-line with how `chempy` works. We'll recreate this by scanning the formulas in `thermo` and then vectorizing the formulas.

thermo_df.loc[10, 'formula']

chlorate = Substance.from_formula(thermo_df.loc[10, 'formula'])
chlorate.composition

for k, v in chlorate.composition.items():
    print (k, v)

Using `chempy` it becomes quite simple to vectorize.

From the composition dictionary:

- key $\Rightarrow$ `stoich[key]`  
- value $\Rightarrow$ `stoich.at[index, key]`

formulas = set(thermo_df['formula'])

stoich_df = pd.DataFrame(formulas)
stoich_df.head()

# naming the column -1 for now so we can sort them later

stoich_df.columns = [-1]
stoich_df.head()

for i, f in enumerate(stoich_df[-1]):
    try:
        sub = Substance.from_formula(f)
        for k, v in sub.composition.items():
            stoich_df.at[i, k] = v
    except:
        pass
        
stoich_df.head()

stoich_df = stoich_df[sorted(stoich_df.columns)]
stoich_df.head()

stoich_df.rename(columns={-1: 'formula'}, inplace=True)
stoich_df.fillna(0, inplace=True)
stoich_df.head()

pickle.dump(stoich_df, open('../data/processed/stoich_df.p', 'wb'))
